Add files via upload
My Data Mining university project in R. This project aimed to generate a classification model to predict the class of an instance (good risk, bad profit or bad loss). The model was generated on the Projet_Data.csv set that was used to build the learning and test sets. Several classifiers were created, and only the most efficient of them was applied.
The selected classifier was then applied to the Projet_Data_New.csv file, which contains new client requests, to predict the class of each of these requests.